ASoC: ops: Fix boolean/integer detection for simple controls
The standard snd_soc_info_volsw() detects if a control is a volume control
and needs to be reported as an integer even if it only has two values by
looking for the string " Volume" in the control name. This results in false
positives if the control has a name like "HP Volume Ramp Switch" since any
" Volume" is matched, not just a trailing one. Fix this by making sure that
we only match at the end of the control name.

    Signed-off-by: Mark Brown <broonie@kernel.org>
    Link: https://lore.kernel.org/r/20220604105407.4055294-1-broonie@kernel.org
    Signed-off-by: Mark Brown <broonie@kernel.org>

Author: Mark Brown <broonie@kernel.org>
Date: Sat Jun 4 11:54:07 2022 +0100

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit aa2a4b897132169fbc6d32932644b95875cf9c7f)
Bugzilla: https://bugzilla.redhat.com/2125537