ASoC: SOF: sof-audio: reset route status before freeing widget
This is in preparation for IPC4 which requires that the route be reset
before the widget is freed. For IPC3, there is nothing more to be done
other than setting the route status. So it is OK to be moved before the
widget is freed.

    Signed-off-by: Ranjani Sridharan <ranjani.sridharan@linux.intel.com>
    Reviewed-by: Péter Ujfalusi <peter.ujfalusi@linux.intel.com>
    Reviewed-by: Bard Liao <yung-chuan.liao@linux.intel.com>
    Reviewed-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>
    Link: https://lore.kernel.org/r/20220426171743.171061-3-ranjani.sridharan@linux.intel.com
    Signed-off-by: Mark Brown <broonie@kernel.org>

Author: Ranjani Sridharan <ranjani.sridharan@linux.intel.com>
Date: Tue Apr 26 10:17:34 2022 -0700

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit 33a3facdf8ccf1b777ef0c39841425ca8d8d4a40)
Bugzilla: https://bugzilla.redhat.com/2125537