ASoC: SOF: ipc4: implement pm ctx_save callback
Use the context save callback to power down the primary core which is used
by the firmware as an indication that the DSP is going to be turned off.

The IMR boot setup is done in response to the primary core power down.

    Signed-off-by: Peter Ujfalusi <peter.ujfalusi@linux.intel.com>
    Reviewed-by: Ranjani Sridharan <ranjani.sridharan@linux.intel.com>
    Reviewed-by: Bard Liao <yung-chuan.liao@linux.intel.com>
    Reviewed-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>
    Link: https://lore.kernel.org/r/20220610083549.16773-7-peter.ujfalusi@linux.intel.com
    Signed-off-by: Mark Brown <broonie@kernel.org>

Author: Peter Ujfalusi <peter.ujfalusi@linux.intel.com>
Date: Fri Jun 10 11:35:49 2022 +0300

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit 63b9069653a710b08d5fd174ac05d43711356541)
Bugzilla: https://bugzilla.redhat.com/2125537