ASoC: audio-graph-card2-custom-sample.dtsi: add verbose explanation
audio-graph-card2-custom-sample.dtsi will be used to test
Audio-Graph-Card2 behavior. But it is difficult to say that it is easy
to understand, because the comment/explanation are not so many.

This patch add verbose explanation to it.

    Signed-off-by: Kuninori Morimoto <kuninori.morimoto.gx@renesas.com>
    Link: https://lore.kernel.org/r/87czeps95h.wl-kuninori.morimoto.gx@renesas.com
    Signed-off-by: Mark Brown <broonie@kernel.org>

Author: Kuninori Morimoto <kuninori.morimoto.gx@renesas.com>
Date: Fri Jul 1 05:18:34 2022 +0000

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit d33083f941150dc2079975d032547f6ce9a8e81b)
Bugzilla: https://bugzilla.redhat.com/2125537