ASoC: adau1761: Add ADAU1761-as-ADAU1361 compatibility mode
During probe, determine if the chip is in fact an ADAU1761
even though an ADAU1361 is specified, and perform additional
operations to enable the ADAU1761 to behave as an ADAU1361,
i.e. disregarding the DSP and setting up routing and PM
transparently.

This enables either chip to be mounted when an ADAU1361 is specified.

    Signed-off-by: Ricard Wanderlof <ricardw@axis.com>
    Link: https://lore.kernel.org/r/alpine.DEB.2.21.2204281841290.5574@lnxricardw1.se.axis.com
    Signed-off-by: Mark Brown <broonie@kernel.org>

Author: Ricard Wanderlof <ricardw@axis.com>
Date: Thu Apr 28 18:46:35 2022 +0200

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit c7b92395836c3962eedbe6b4f58f9dfb2b2cc921)
Bugzilla: https://bugzilla.redhat.com/2125537