ASoC: cs35l41: Do not print error when waking from hibernation
When waking from hibernation, it is possible for the function
which sends the wake command to fail initially, but after a
retry it will succeed. There is no need to print an error if
the initial attempts fail.

    Signed-off-by: Stefan Binding <sbinding@opensource.cirrus.com>
    Signed-off-by: Vitaly Rodionov <vitalyr@opensource.cirrus.com>
    Acked-by: Charles Keepax <ckeepax@opensource.cirrus.com>
    Link: https://lore.kernel.org/r/20220525131638.5512-11-vitalyr@opensource.cirrus.com
    Signed-off-by: Mark Brown <broonie@kernel.org>

Author: Stefan Binding <sbinding@opensource.cirrus.com>
Date: Wed May 25 14:16:31 2022 +0100

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit 97076475e2fdf471348b9ce73215cdbceeb4390f)
Bugzilla: https://bugzilla.redhat.com/2125537