ASoC: tas2770: Reinit regcache on reset
On probe of the ASoC component, the device is reset but the regcache is
retained. This means the regcache gets out of sync if the codec is
rebound to a sound card for a second time. Fix it by reinitializing the
regcache to defaults after the device is reset.

    Fixes: b0bcbe615756 ("ASoC: tas2770: Fix calling reset in probe")
    Signed-off-by: Martin Povišer <povik+lin@cutebit.org>
    Link: https://lore.kernel.org/r/20220919173453.84292-1-povik+lin@cutebit.org
    Signed-off-by: Mark Brown <broonie@kernel.org>

Author: Martin Povišer <povik+lin@cutebit.org>
Date: Mon Sep 19 19:34:53 2022 +0200

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit 0a0342ede303fc420f3a388e1ae82da3ae8ff6bd)
Bugzilla: https://bugzilla.redhat.com/2125537