ASoC: soc-utils: Add kunit test for snd_soc_tdm_params_to_bclk()
Create a new kunit test for soc-utils and use it to test
snd_soc_tdm_params_to_bclk().

The test uses a table of values to avoid the possibility that an
on-the-fly generator contains the same algorithmic error as the
function-under-test and so fails to detect a bug.

There is no need to test every possible combination of values.
Enough test cases are included to give confidence that the function
is producing the correct results.

    Signed-off-by: Richard Fitzgerald <rf@opensource.cirrus.com>
    Link: https://lore.kernel.org/r/20220405135419.1230088-4-rf@opensource.cirrus.com
    Signed-off-by: Mark Brown <broonie@kernel.org>

Author: Richard Fitzgerald <rf@opensource.cirrus.com>
Date: Tue Apr 5 14:54:17 2022 +0100

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit 89342fa38bbaade51584f255eee5cd43621f4e10)
Bugzilla: https://bugzilla.redhat.com/2125537