ASoC: fsl_micfil: rework quality setting
For the quality setting the quality setting register values are directly
exposed to the kcontrol and thus to userspace. This is unfortunate
because the register settings contains invalid bit combinations marked
as "N/A". For userspace it doesn't make much sense to be able to set
these just to see that the driver responds with "Please make sure you
select a valid quality." in the kernel log.

Work around this by adding get/set functions for the quality setting.

    Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>
    Acked-by: Shengjiu Wang <shengjiu.wang@gmail.com>
    Link: https://lore.kernel.org/r/20220414162249.3934543-17-s.hauer@pengutronix.de
    Signed-off-by: Mark Brown <broonie@kernel.org>

Author: Sascha Hauer <s.hauer@pengutronix.de>
Date: Thu Apr 14 18:22:44 2022 +0200

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit bea1d61d5892083551f92818e964801baabd95a9)
Bugzilla: https://bugzilla.redhat.com/2125537