ASoC: fsl_sai: add error message in case of missing imx-pcm-dma support
If the imx-pcm-dma is required we need to have the module enabled. For
all NXP/FSL sound cards using the ASoC architecture this is the case but
in case of using the simple-audio-card sound card this isn't the case.

In such case the driver probe fails silently and the card isn't
available. It took a while to find the missing Kconfig. Make this easier
for others by printing a error if this the module isn't available but
required by the HW.

    Signed-off-by: Marco Felsch <m.felsch@pengutronix.de>
    Acked-by: Shengjiu Wang <shengjiu.wang@gmail.com>
    Link: https://lore.kernel.org/r/20220601092342.3328644-3-m.felsch@pengutronix.de
    Signed-off-by: Mark Brown <broonie@kernel.org>

Author: Marco Felsch <m.felsch@pengutronix.de>
Date: Wed Jun 1 11:23:42 2022 +0200

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit 22205521770ee740f64a3ec90301f50e34738cfd)
Bugzilla: https://bugzilla.redhat.com/2125537