OPP: Add dev_pm_opp_set_config() and friends
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=2122311

commit 11b9b663585c4f8b00846089ebbca4d1e3283e86
Author: Viresh Kumar <viresh.kumar@linaro.org>
Date:   Wed May 25 15:23:16 2022 +0530

The OPP core already have few configuration specific APIs and it is
getting complex or messy for both the OPP core and its users.

Lets introduce a new set of API which will be used for all kind of
different configurations, and shall eventually be used by all the
existing ones.

The new API, returns a unique token instead of a pointer to the OPP
table, which allows the OPP core to drop the resources selectively later
on.

Tested-by: Dmitry Osipenko <dmitry.osipenko@collabora.com>
Signed-off-by: Viresh Kumar <viresh.kumar@linaro.org>
Signed-off-by: Mark Langsdorf <mlangsdo@redhat.com>