OPP: Migrate set-regulators API to use set-config helpers
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=2122311

commit b0ec09428621daee5101130c307634a390b0213b
Author: Viresh Kumar <viresh.kumar@linaro.org>
Date: Mon, 4 Jul 2022 16:36:26 +0530

Now that we have a central API to handle all OPP table configurations,
migrate the set-regulators family of helpers to use the new
infrastructure.

The return type and parameter to the APIs change a bit due to this,
update the current users as well in the same commit in order to avoid
breaking builds.

Reviewed-by: Chanwoo Choi <cw00.choi@samsung.com>
Signed-off-by: Viresh Kumar <viresh.kumar@linaro.org>
Signed-off-by: Mark Langsdorf <mlangsdo@redhat.com>