OPP: Migrate set-supported-hw API to use set-config helpers
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=2122311

commit 89f03984fa2abface1ffb1fe050b7c175651ffc7
Author: Viresh Kumar <viresh.kumar@linaro.org>
Date: Thu, 26 May 2022 09:36:27 +0530

Now that we have a central API to handle all OPP table configurations,
migrate the set-supported-hw family of helpers to use the new
infrastructure.

The return type and parameter to the APIs change a bit due to this,
update the current users as well in the same commit in order to avoid
breaking builds.

Signed-off-by: Viresh Kumar <viresh.kumar@linaro.org>
Signed-off-by: Mark Langsdorf <mlangsdo@redhat.com>