docs: mention that deps via MR number work too
We've had support in the webhooks for a while now for specifying
dependencies via dependent MR number -- either full URL to the MR or the
gitlab !-prefixed shorthand. Under the covers, we look up the bug number
from the dependency MR's description, and it's much closer to what
GitLab's built-in (but as yet insufficient) dependency tracking uses.
We'll refrain from documenting the !x shorthand for now though, and may
consider removing it, due to it's limited usefulness if using cli.

Signed-off-by: Jarod Wilson <jarod@redhat.com>