perf/x86/amd/uncore: Use dynamic events array
Bugzilla: https://bugzilla.redhat.com/2123229

upstream
========
commit 39621c5808f5dda75d03dc4b2d4d2b13a5a1c34b
Author: Sandipan Das <sandipan.das@amd.com>
Date: Thu May 19 15:33:30 2022 +0530

description
===========
If AMD Performance Monitoring Version 2 (PerfMonV2) is
supported, the number of available counters for a given
uncore PMU may not be fixed across families and models
and has to be determined at runtime.

The per-cpu uncore PMU data currently uses a fixed-sized
array for event information. Make it dynamic based on the
number of available counters.

    Signed-off-by: Sandipan Das <sandipan.das@amd.com>
    Signed-off-by: Peter Zijlstra (Intel) <peterz@infradead.org>
    Link: https://lore.kernel.org/r/21eea0cb6de9d14f78d52d1d62637ae02bc900f5.1652954372.git.sandipan.das@amd.com

Signed-off-by: Michael Petlan <mpetlan@redhat.com>