perf/x86/intel: Fix PEBS memory access info encoding for ADL
Bugzilla: https://bugzilla.redhat.com/2123229

upstream
========
commit 39a41278f041e4b7ee6c83caefac845c9b19fc61
Author: Kan Liang <kan.liang@linux.intel.com>
Date: Wed Jun 29 08:08:39 2022 -0700

description
===========
The PEBS memory access latency encoding for the e-core is slightly
different from the p-core. The bit 4 is Lock, while the bit 5 is TLB
access.

Add a new flag to indicate the load/store latency event on a hybrid
platform.
Add a new function pointer to retrieve the latency data for a hybrid
platform. Only implement the new flag and function for the e-core on
ADL. Still use the existing PERF_X86_EVENT_PEBS_LDLAT/STLAT flag for the
p-core on ADL.

Factor out pebs_set_tlb_lock() to set the generic memory data source
information of the TLB access and lock for both load and store latency.

Move the intel_get_event_constraints() to ahead of the :ppp check,
otherwise the new flag never gets a chance to be set for the :ppp
events.

Fixes: f83d2f91d259 ("perf/x86/intel: Add Alder Lake Hybrid support")
    Signed-off-by: Kan Liang <kan.liang@linux.intel.com>
    Signed-off-by: Peter Zijlstra (Intel) <peterz@infradead.org>
    Reviewed-by: Andi Kleen <ak@linux.intel.com>
    Link: https://lkml.kernel.org/r/20220629150840.2235741-1-kan.liang@linux.intel.com

Signed-off-by: Michael Petlan <mpetlan@redhat.com>