perf record: Fix manpage formatting of description of support to hybrid systems
Bugzilla: https://bugzilla.redhat.com/2123229

upstream
========
commit e89eaa611c7568d1288a2ccca88355a9434f2d47
Author: Andi Kleen <ak@linux.intel.com>
Date: Thu Aug 18 03:01:27 2022 -0700

description
===========
The Intel hybrid description is written in a different style than the
rest of the perf record man page. There were some new command line
options added after it which resulted in very strange section ordering.
Move the hybrid include last.

Also the sub sections in the hybrid document don't fit the record
manpage well (especially since it talks about all kinds of unrelated
commands). I left this for now, but would be better to separate this
properly in the different man pages.

It would be better to use sub sections for the other sections, but these
don't seem to be supported in AsciiDoc?

Some of the examples are still misrendered in the manpage with an
indented troff command, but I don't know how to fix that.

In any case it's now better than before.

    Signed-off-by: Andi Kleen <ak@linux.intel.com>
    Cc: zhengjun.xing@intel.com
    Link: https://lore.kernel.org/r/20220818100127.249401-1-ak@linux.intel.com
    Signed-off-by: Arnaldo Carvalho de Melo <acme@redhat.com>

Signed-off-by: Michael Petlan <mpetlan@redhat.com>