cifs: track individual channel status using chans_need_reconnect
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=2138995
commit d1a931ce2e3b7761d293ba8e0bde2b0180f456e9

We needed a way to identify the channels under the smb session
which are in reconnect, so that the traffic to other channels
can continue. So I replaced the bool need_reconnect with
a bitmask identifying all the channels that need reconnection
(named chans_need_reconnect). When a channel needs reconnection,
the bit corresponding to the index of the server in ses->chans
is used to set this bitmask. Checking if no channels or all
the channels need reconnect then becomes very easy.

Also wrote some helper macros for checking and setting the bits.

    Signed-off-by: Shyam Prasad N <sprasad@microsoft.com>
    Signed-off-by: Steve French <stfrench@microsoft.com>
    (cherry picked from commit d1a931ce2e3b7761d293ba8e0bde2b0180f456e9)
Signed-off-by: Ronnie Sahlberg <lsahlber@redhat.com>