cifs: use the chans_need_reconnect bitmap for reconnect status
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=2138995
commit f486ef8e2003f6c308d0db81ea116c880a760d4f

We use the concept of "binding" when one of the secondary channel
is in the process of connecting/reconnecting to the server. Till this
binding process completes, and the channel is bound to an existing session,
we redirect traffic from other established channels on the binding channel,
effectively blocking all traffic till individual channels get reconnected.

With my last set of commits, we can get rid of this binding serialization.
We now have a bitmap of connection states for each channel. We will use
this bitmap instead for tracking channel status.

Having a bitmap also now enables us to keep the session alive, as long
as even a single channel underneath is alive.

Unfortunately, this also meant that we need to supply the tcp connection
info for the channel during all negotiate and session setup functions.
These changes have resulted in a slightly bigger code churn.
However, I expect perf and robustness improvements in the mchan scenario
after this change.

    Signed-off-by: Shyam Prasad N <sprasad@microsoft.com>
    Signed-off-by: Steve French <stfrench@microsoft.com>
    (cherry picked from commit f486ef8e2003f6c308d0db81ea116c880a760d4f)
Signed-off-by: Ronnie Sahlberg <lsahlber@redhat.com>