cifs: reconnect only the connection and not smb session where possible
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=2138995
commit 183eea2ee5ba968ca7c31f04a0f01fd3e5c1d014

With the new per-channel bitmask for reconnect, we have an option to
reconnect the tcp session associated with the channel without reconnecting
the smb session. i.e. if there are still channels to operate on, we can
continue to use the smb session and tcon.

However, there are cases where it makes sense to reconnect the smb session
even when there are active channels underneath. For example for
SMB session expiry.

With this patch, we'll have an option to do either, and use the correct
option for specific cases.

    Signed-off-by: Shyam Prasad N <sprasad@microsoft.com>
    Signed-off-by: Steve French <stfrench@microsoft.com>
    (cherry picked from commit 183eea2ee5ba968ca7c31f04a0f01fd3e5c1d014)
Signed-off-by: Ronnie Sahlberg <lsahlber@redhat.com>