cifs: protect all accesses to chan_* with chan_lock
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=2138995
commit 88b024f556fcd5bf1288c6333016f576cfa5f539

A spin lock called chan_lock was introduced recently.
But not all accesses were protected. Doing that with
this change.

To make sure that a channel is not freed when in use,
we need to introduce a ref count. But today, we don't
ever free channels.

    Signed-off-by: Shyam Prasad N <sprasad@microsoft.com>
    Signed-off-by: Steve French <stfrench@microsoft.com>
    (cherry picked from commit 88b024f556fcd5bf1288c6333016f576cfa5f539)
Signed-off-by: Ronnie Sahlberg <lsahlber@redhat.com>