cifs: mark sessions for reconnection in helper function
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=2138995
commit 2a05137a0575b7d1006bdf4c1beeee9e391e22a0

Today we have the code to mark connections and sessions
(and tcons) for reconnect clubbed with the code to close
the socket and abort all mids in the same function.

Sometimes, we need to mark connections and sessions
outside cifsd thread. So as a part of this change, I'm
splitting this function into two different functions and
calling them one after the other in cifs_reconnect.

    Signed-off-by: Shyam Prasad N <sprasad@microsoft.com>
    Signed-off-by: Steve French <stfrench@microsoft.com>
    (cherry picked from commit 2a05137a0575b7d1006bdf4c1beeee9e391e22a0)
Signed-off-by: Ronnie Sahlberg <lsahlber@redhat.com>