cifs: remove check of list iterator against head past the loop body
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=2138995
commit 00c796eecba4898194ea549679797ee28f89a92f

When list_for_each_entry() completes the iteration over the whole list
without breaking the loop, the iterator value will be a bogus pointer
computed based on the head element.

While it is safe to use the pointer to determine if it was computed
based on the head element, either with list_entry_is_head() or
&pos->member == head, using the iterator variable after the loop should
be avoided.

In preparation to limit the scope of a list iterator to the list
traversal loop, use a dedicated pointer to point to the found element [1].

    Link: https://lore.kernel.org/all/CAHk-=wgRr_D8CB-D9Kg-c=EHreAsk5SqXPwr9Y7k9sA6cWXJ6w@mail.gmail.com/ [1]
    Reviewed-by: Paulo Alcantara (SUSE) <pc@cjr.nz>
    Signed-off-by: Jakob Koschel <jakobkoschel@gmail.com>
    Signed-off-by: Steve French <stfrench@microsoft.com>
    (cherry picked from commit 00c796eecba4898194ea549679797ee28f89a92f)
Signed-off-by: Ronnie Sahlberg <lsahlber@redhat.com>