cifs: Split the smb3_add_credits tracepoint
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=2138995
commit 1ddff774164f1bcd0fcf988f7a5bb24270fbdf2c

Split the smb3_add_credits tracepoint to make it more obvious when looking
at the logs which line corresponds to what credit change.  Also add a
tracepoint for credit overflow when it's being added back.

Note that it might be better to add another field to the tracepoint for
the information rather than splitting it.  It would also be useful to store
the MID potentially, though that isn't available when the credits are first
obtained.

    Signed-off-by: David Howells <dhowells@redhat.com>
    cc: Shyam Prasad N <nspmangalore@gmail.com>
    cc: Rohith Surabattula <rohiths.msft@gmail.com>
    cc: linux-cifs@vger.kernel.org
    Acked-by: Paulo Alcantara (SUSE) <pc@cjr.nz>
    Reviewed-by: Enzo Matsumiya <ematsumiya@suse.de>
    Signed-off-by: Steve French <stfrench@microsoft.com>
    (cherry picked from commit 1ddff774164f1bcd0fcf988f7a5bb24270fbdf2c)
Signed-off-by: Ronnie Sahlberg <lsahlber@redhat.com>