cifs: check for smb1 in open_cached_dir()
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=2138995
commit 198bf836dfb9c76105c93e6a3138ed1a33841fad

Check protocol version in open_cached_dir() and return not supported
for SMB1.  This allows us to call open_cached_dir() from code that
is common to both smb1 and smb2/3 in future patches without having to
do this check in the call-site.
At the same time, add a check if tcon is valid or not for the same reason.

    Reviewed-by: Paulo Alcantara (SUSE) <pc@cjr.nz>
    Reviewed-by: Enzo Matsumiya <ematsumiya@suse.de>
    Signed-off-by: Ronnie Sahlberg <lsahlber@redhat.com>
    Signed-off-by: Steve French <stfrench@microsoft.com>
    (cherry picked from commit 198bf836dfb9c76105c93e6a3138ed1a33841fad)
Signed-off-by: Ronnie Sahlberg <lsahlber@redhat.com>