cifs: set the CREATE_NOT_FILE when opening the directory in use_cached_dir()
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=2138995
commit c9fc5ca454b3f0606589bff306793a3d2f0d9f13

This enforces that we can only do this for directories and not normal files
or else the server will return an error.
This means that we will have conditionally check IF the path refers
to a directory or not in all the call-sites where we are unsure.
Right now this check is for "" i.e. root.

    Reviewed-by: Paulo Alcantara (SUSE) <pc@cjr.nz>
    Reviewed-by: Enzo Matsumiya <ematsumiya@suse.de>
    Signed-off-by: Ronnie Sahlberg <lsahlber@redhat.com>
    Signed-off-by: Steve French <stfrench@microsoft.com>
    (cherry picked from commit c9fc5ca454b3f0606589bff306793a3d2f0d9f13)
Signed-off-by: Ronnie Sahlberg <lsahlber@redhat.com>