smb2: small refactor in smb2_check_message()
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=2138995
commit da3847894fddc27ca95d5ac0012f444a77a5e0c3

If the command is SMB2_IOCTL, OutputLength and OutputContext are
optional and can be zero, so return early and skip calculated length
check.

Move the mismatched length message to the end of the check, to avoid
unnecessary logs when the check was not a real miscalculation.

Also change the pr_warn_once() to a pr_warn() so we're sure to get a
log for the real mismatches.

    Signed-off-by: Enzo Matsumiya <ematsumiya@suse.de>
    Signed-off-by: Steve French <stfrench@microsoft.com>
    (cherry picked from commit da3847894fddc27ca95d5ac0012f444a77a5e0c3)
Signed-off-by: Ronnie Sahlberg <lsahlber@redhat.com>