cpuidle: Add cpu_idle_miss trace event
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=2122313

commit 6ab4b1990097b76508bd6dffd85ffcacbedb26b2
Author: Kajetan Puchalski <kajetan.puchalski@arm.com>
Date: Tue, 26 Jul 2022 11:24:04 +0100

Add a trace event for cpuidle to track missed (too deep or too shallow)
wakeups.

After each wakeup, CPUIdle already computes whether the entered state was
optimal, above or below the desired one and updates the relevant
counters. This patch makes it possible to trace those events in addition
to just reading the counters.

The patterns of types and percentages of misses across different
workloads appear to be very consistent. This makes the trace event very
useful for comparing the relative correctness of different CPUIdle
governors for different types of workloads, or for finding the
optimal governor for a given device.

Signed-off-by: Kajetan Puchalski <kajetan.puchalski@arm.com>
Reviewed-by: Steven Rostedt (Google) <rostedt@goodmis.org>
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Signed-off-by: Mark Langsdorf <mlangsdo@redhat.com>