redhat: Add sub-RPM with a EFI unified kernel image for virtual machines
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=2142102
Upstream Status: RHEL only

The new 'kernel-unified-virt' sub-RPM is added on x86_64 targets.

This contains an EFI application that provides a combined vmlinux,
initrd and cmdline, as a so called 'unified kernel image'. The
spec for this is defined by the boot loader specification

  https://uapi-group.org/specifications/specs/boot_loader_specification/

The key benefit of a unified kernel is that its secure boot
signature covers the initrd and cmdline contents, allowing
a trustworthy measured boot process with attestation, which
is not practical with locally generated initrds/cmdlines.

Since the initrd is pre-generated its contents have to be
very generic, to be usable on a wide variety of deployments.
To make this problem tractable, the sub-RPM targets only
usage in virtual machines. With such a restriction, the
initrd only needs a very small set of block driver modules
present, in order to be usable across KVM, Hyper-V and Xen
hypervisors which will cover essentially all common public
and private clouds.

Similarly the kernel cmdline cannot contain any host specific
data, which means the root filesystem to mount needs to be
able to be automatically detected. A virtual machine image
intending to use this unified kernel package thus needs to
comply with the discoverable partitions specification:

  https://uapi-group.org/specifications/specs/discoverable_partitions_specification/

Based-on-patch-by: Daniel P. Berrangé <berrange@redhat.com>
Based-on-patch-by: Gerd Hoffmann <kraxel@redhat.com>
Signed-off-by: Vitaly Kuznetsov <vkuznets@redhat.com>