platform/chrome: cros_ec_proto: Rename cros_ec_command function
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=2122315
Conflicts:
	drivers/mfd/cros_ec_dev.c, drivers/platform/chrome/cros_ec_proto.c,
drivers/platform/chrome/cros_ec_typec.c, drivers/platform/chrome/cros_usbpd_notify.c,
drivers/regulator/cros-ec-regulator.c, include/linux/platform_data/cros_ec_proto.h -
Not all of the files had all the instances of the cros_ec_command() function to
rename, and none of the function had extra instances of it. Rename the function
when it was present and do nothing when it was not.

commit b1d288d9c3c5ca28df062214656a59cf7ee370e0
Author: Prashant Malani <pmalani@chromium.org>
Date: Mon, 6 Jun 2022 20:18:03 +0000

cros_ec_command() is the name of a function as well as a struct, as such
it can confuse indexing tools (like ctags). Avoid this by renaming it to
cros_ec_cmd(). Update all the callsites to use the new name.

This patch is a find-and-replace, so should not introduce any functional
changes.

Suggested-by: Stephen Boyd <swboyd@chromium.org>
Signed-off-by: Prashant Malani <pmalani@chromium.org>
Acked-by: Lee Jones <lee.jones@linaro.org>
Reviewed-by: Stephen Boyd <swboyd@chromium.org>
Reviewed-by: Guenter Roeck <groeck@chromium.org>
Signed-off-by: Tzung-Bi Shih <tzungbi@kernel.org>
Link: https://lore.kernel.org/r/20220606201825.763788-3-pmalani@chromium.org
Signed-off-by: Mark Langsdorf <mlangsdo@redhat.com>